Moved AvailabilityScreen.kt from android/sample to github.se.project. Deleted android/sample directory and modified main to display the availability screen.
