+++ app/src/main/java/com/github/se/project/ui/authentication/AvailabilityScreen.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.authentication
+package com.github.se.project.ui.authentication
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -18,10 +18,11 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.android.sample.model.profile.AcademicLevel
-import com.android.sample.model.profile.Profile
-import com.android.sample.model.profile.Role
-import com.android.sample.model.profile.Section
+import com.github.se.project.model.profile.Profile
+import com.github.se.project.model.profile.AcademicLevel
+import com.github.se.project.model.profile.Role
+import com.github.se.project.model.profile.Section
+
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
